+++ BE/src/InitDB/InitDB.module.ts
@@ -3,10 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserModel } from '../user/entities/user.entity';
 import { UserQuizArchiveModel } from '../user/entities/user-quiz-archive.entity';
 import { InitDBService } from './InitDB.Service';
-import { QuizModel } from '../quiz/entities/quiz.entity';
-import { QuizSetModel } from '../quiz/entities/quiz-set.entity';
+import { QuizModel } from '../quiz-set/entities/quiz.entity';
+import { QuizSetModel } from '../quiz-set/entities/quiz-set.entity';
 import { InitDBController } from './InitDB.controller';
-import { QuizChoiceModel } from '../quiz/entities/quiz-choice.entity';
+import { QuizChoiceModel } from '../quiz-set/entities/quiz-choice.entity';
 
 @Module({
   imports: [
