+++ BE/src/app.module.ts
@@ -5,14 +5,14 @@ import { GameModule } from './game/game.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { RedisModule } from '@nestjs-modules/ioredis';
 import { ConfigModule } from '@nestjs/config';
-import { QuizSetModel } from './quiz/entities/quiz-set.entity';
-import { QuizModel } from './quiz/entities/quiz.entity';
-import { QuizChoiceModel } from './quiz/entities/quiz-choice.entity';
+import { QuizSetModel } from './quiz-set/entities/quiz-set.entity';
+import { QuizModel } from './quiz-set/entities/quiz.entity';
+import { QuizChoiceModel } from './quiz-set/entities/quiz-choice.entity';
 import { UserModel } from './user/entities/user.entity';
 import { UserQuizArchiveModel } from './user/entities/user-quiz-archive.entity';
 import { InitDBModule } from './InitDB/InitDB.module';
 import { UserModule } from './user/user.module';
-import { QuizModule } from './quiz/quiz.module';
+import { QuizSetModule } from './quiz-set/quiz-set.module';
 import { WaitingRoomModule } from './waiting-room/waiting-room.module';
 
 @Module({
@@ -42,7 +42,7 @@ import { WaitingRoomModule } from './waiting-room/waiting-room.module';
       type: 'single',
       url: process.env.REDIS_URL || 'redis://localhost:6379'
     }),
-    QuizModule,
+    QuizSetModule,
     UserModule,
     InitDBModule,
     WaitingRoomModule
